[Design] #22 - 내 짐 레이아웃 구현
