+++ PickPack/Models/Item.swift
@@ -21,8 +21,6 @@ struct Item: Identifiable, Codable, Hashable {
     var isPacked: Bool = false
     
     var quantity: Int = 1
-    
-    var state: itemState = .normal
 }
 
 enum ItemTheme: String, Codable, CaseIterable, Hashable {
@@ -38,14 +36,4 @@ enum ItemTheme: String, Codable, CaseIterable, Hashable {
     case golfGear = "골프용품"
     case seasonalItems = "계절용품"
     case others = "기타"
-    
-    // MyItemView에서 제일 처음 그려지는 테마에서는 선을 그려주지 않기 위해 작성되었습니다
-    static func firstMatchingTheme(items: [Item]) -> ItemTheme? {
-        for theme in ItemTheme.allCases {
-            if !items.filter({ $0.theme == theme }).isEmpty {
-                return theme
-            }
-        }
-        return nil
-    }
 }
